requirements.txt
python-telegram-bot==20.7
python-jobspy
joblib
requests
beautifulsoup4
pandas
google-generativeai==0.3.1
pdfplumber==0.10.3
python-docx==1.1.0
python-dotenv==1.0.1
sentence-transformers
pdfplumber
plotly
